Hello AoC 2022
Solution for day 1 of previous AoC is added as a reference for testing
reasons and for Thursday reference since I'm rusty in Rust (haha).

Fetching puzzle input is automated if there is AOC_SESSION env var set
with the, well, AoC session cookie.
If env var is not set, input has to be manually saved into input/day1.txt etc.