Add Day 22 part 2 (non-generic) solution
Holy shit, this one is a story to remember
This was meant to be a generic solution (and can be very easily made
into one!), but I'm out of energy and time for philosophical Paint
drawings.
I'm super happy it works for both my input and example input.
To support more grids, the rest of cases has to be added to dissect_cube
function. Maybe someday (haha)